requirements.txt
Mastodon.py>=1.5.2
toml>=0.10.2